+++ examples/beginners_diffusion/MMSPDiffusion2D.cpp
@@ -19,8 +19,8 @@ int main()
 	std::cin>>iterate;
 	std::cout<<""<<std::endl;
 
-	grid<2,scalar<float> > GRID(1,0,length,0,length);
-	grid<2,scalar<float> > update(1,0,length,0,length);
+	grid<2,scalar<double> > GRID(1,0,length,0,length);
+	grid<2,scalar<double> > update(1,0,length,0,length);
 
 	for (int x=x0(GRID); x<x1(GRID); x++) {
 		for (int y=y0(GRID); y<y1(GRID); y++) {
